// Atrasti.ReactNative/Atrasti/src/Api/Models/Responds/AgentPage_Res.ts
export interface AgentPage_Res {
    logo: string | null;
    address: string;
    phoneNumber: string;
    country: string;
    county: string;
    city: string;
    website: string;
    description: string;
    businessSector: string;
    mainProducts: string;
    mainMarkets: string;
    yearsOfExperience: string;
    certificates: string;
}