// Atrasti.ReactNative/Atrasti/src/Api/Models/Requests/ManageAgent_Req.ts
export interface ManageAgent_Req {
    logo: string;
    address: string;
    phoneNumber: string;
    country: string;
    county: string;
    city: string;
    website: string;
    description: string;
    businessSector: string;
    mainProducts: string;
    mainMarkets: string;
    yearsOfExperience: string;
    certificates: string;
}